[1129] 引入 goldfish package,移除 s7 package
在 xmake/goldfish.lua 新增 package("goldfish") 定义,源码来自
TeXmacs/plugins/goldfish/src/,on_install 在包缓存目录内联生成
libgoldfish 静态库 target 并公开 s7.h。libmogan/stem/tests 改用
add_packages("goldfish"),删除 xmake/packages/s/s7/。

goldfish binary target 保持不变。
