+++ xmake/tests.lua
@@ -29,7 +29,7 @@ function add_target_cpp_test(filepath, dep1, dep2)
         if not is_plat("windows") then
             add_syslinks("pthread")
         end
-        add_packages("s7")
+        add_packages("goldfish")
         add_packages("liii-pdfhummus")
 
         add_includedirs({"$(builddir)", "tests/Base"})
@@ -78,7 +78,7 @@ function add_target_cpp_bench(filepath, dep)
         if not is_plat("windows") then
             add_syslinks("pthread")
         end
-        add_packages("s7")
+        add_packages("goldfish")
         add_packages("liii-pdfhummus")
 
         add_includedirs({"$(builddir)", "tests/Base"})
